# docs/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Chess Alive API",
...
    "version": "1.0.0",
    "contact": {
      "name": "Chess Alive Team",
      "url": "https://github.com/yourusername/chess-alive"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Development server"
    }
  ],
...
  "paths": {
...
    "/readiness": {
      "get": {
        "tags": ["health"],
        "summary": "Readiness Check",
        "description": "Kubernetes/Cloud Run readiness probe. Returns 503 if not ready to serve traffic.",
        "responses": {
          "200": {
            "description": "Ready",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadinessCheck"
                }
              }
            }
          },
          "503": {
            "description": "Not Ready"
          }
        }
      }
    },
...
    "/api/v1/games": {
      "post": {
        "tags": ["games"],
        "summary": "Create Game",
        "description": "Create a new game. Returns a share code that other players can use to join.",
        "parameters": [
          {
            "name": "X-Session-ID",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Player session ID (auto-generated if not provided)"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGameRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Game created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/games/join-by-code": {
      "post": {
        "tags": ["games"],
        "summary": "Join Game by Code",
        "description": "Join an existing game using its share code.",
        "parameters": [
          {
            "name": "X-Session-ID",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JoinGameRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Joined game successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameResponse"
                }
              }
            }
          },
          "404": {
            "description": "Invalid share code"
          }
        }
      }
    },
...
    "/api/v1/games/{game_id}/command": {
      "post": {
        "tags": ["game-actions"],
        "summary": "Issue Command",
        "description": "Issue a move command to a piece. The piece may refuse based on its morale and the riskiness of the move.",
        "parameters": [
          {
            "name": "game_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Session-ID",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommandRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Command processed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid move or not your turn"
          },
          "404": {
            "description": "Game or piece not found"
          }
        }
      }
    },
    "/api/v1/games/{game_id}/persuade": {
      "post": {
        "tags": ["game-actions"],
        "summary": "Persuade Piece",
        "description": "Submit a persuasion attempt when a piece refuses to move. The AI evaluates your argument.",
        "parameters": [
          {
            "name": "game_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Session-ID",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PersuasionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Persuasion processed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersuasionResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/games/{game_id}/chat": {
      "get": {
        "tags": ["chat"],
        "summary": "Get Chat History",
        "description": "Get paginated chat history for a game.",
        "parameters": [
          {
            "name": "game_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Chat messages",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ChatMessage"
                      }
                    },
                    "total": {"type": "integer"},
                    "page": {"type": "integer"},
                    "page_size": {"type": "integer"},
                    "has_more": {"type": "boolean"}
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["chat"],
        "summary": "Send Message",
        "description": "Send a message in the game chat.",
        "parameters": [
          {
            "name": "game_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Session-ID",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendMessageRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Message sent"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ReadinessCheck": {
        "type": "object",
        "properties": {
          "status": {"type": "string", "example": "ready"}
        }
      },
...
      "CreateGameRequest": {
        "type": "object",
        "properties": {
          "game_mode": {"type": "string", "enum": ["pvp", "pvai"], "default": "pvp"},
          "template": {"type": "string", "enum": ["classic", "power_chess", "leaper_madness", "hopper_havoc", "pawn_revolution"], "default": "classic"},
          "settings": {"$ref": "#/components/schemas/GameSettings"}
        }
      },
      "GameSettings": {
        "type": "object",
        "properties": {
          "surprise_mode": {"type": "boolean", "default": false},
          "turn_timer": {"type": "integer", "nullable": true},
          "ai_difficulty": {"type": "string", "enum": ["easy", "medium", "hard"], "default": "medium"}
        }
      },
      "JoinGameRequest": {
        "type": "object",
        "required": ["share_code"],
        "properties": {
          "share_code": {"type": "string", "pattern": "^[A-Z0-9]{6,12}$"}
        }
      },
      "GameResponse": {
        "type": "object",
        "properties": {
          "id": {"type": "string"},
          "status": {"type": "string", "enum": ["waiting", "active", "completed"]},
          "game_mode": {"type": "string"},
          "template": {"type": "string"},
          "share_code": {"type": "string"},
          "fen": {"type": "string"},
          "turn": {"type": "string", "enum": ["white", "black"]},
          "white_player": {"$ref": "#/components/schemas/PlayerSummary"},
          "black_player": {"$ref": "#/components/schemas/PlayerSummary"},
          "pieces": {"type": "array", "items": {"$ref": "#/components/schemas/PieceState"}},
          "settings": {"$ref": "#/components/schemas/GameSettings"},
          "created_at": {"type": "string", "format": "date-time"}
        }
      },
      "PlayerSummary": {
        "type": "object",
        "properties": {
          "id": {"type": "string"},
          "username": {"type": "string"},
          "avatar_url": {"type": "string", "nullable": true}
        }
      },
      "PieceState": {
        "type": "object",
        "properties": {
          "id": {"type": "string"},
          "color": {"type": "string", "enum": ["white", "black"]},
          "piece_type": {"type": "string", "enum": ["pawn", "knight", "bishop", "rook", "queen", "king"]},
          "square": {"type": "string", "nullable": true},
          "morale": {"type": "integer", "minimum": 0, "maximum": 100, "default": 70},
          "personality": {"$ref": "#/components/schemas/PiecePersonality"},
          "is_captured": {"type": "boolean", "default": false}
        }
      },
      "PiecePersonality": {
        "type": "object",
        "properties": {
          "archetype": {"type": "string", "default": "default"},
          "traits": {"type": "array", "items": {"type": "string"}},
          "dialogue_style": {"type": "string", "default": "neutral"},
          "custom_prompt": {"type": "string", "default": ""},
          "morale_modifiers": {"type": "object"}
        }
      },
      "CommandRequest": {
        "type": "object",
        "required": ["piece_id", "target_square"],
        "properties": {
          "piece_id": {"type": "string"},
          "target_square": {"type": "string", "pattern": "^[a-h][1-8]$"},
          "message": {"type": "string", "maxLength": 500}
        }
      },
      "CommandResponse": {
        "type": "object",
        "properties": {
          "success": {"type": "boolean"},
          "move_executed": {"type": "boolean"},
          "piece_response": {"$ref": "#/components/schemas/PieceResponseData"},
          "board_state": {"$ref": "#/components/schemas/BoardState"},
          "morale_changes": {"type": "array", "items": {"$ref": "#/components/schemas/MoraleChange"}},
          "analysis": {"$ref": "#/components/schemas/MoveAnalysis"},
          "king_taunt": {"type": "string", "nullable": true}
        }
      },
      "PieceResponseData": {
        "type": "object",
        "properties": {
          "will_move": {"type": "boolean"},
          "response_text": {"type": "string"},
          "morale_before": {"type": "integer"},
          "morale_after": {"type": "integer"},
          "reason_for_refusal": {"type": "string", "nullable": true}
        }
      },
      "BoardState": {
        "type": "object",
        "properties": {
          "fen": {"type": "string"},
          "turn": {"type": "string"},
          "is_check": {"type": "boolean", "default": false},
          "is_checkmate": {"type": "boolean", "default": false},
          "is_stalemate": {"type": "boolean", "default": false},
          "last_move": {"$ref": "#/components/schemas/MoveData"}
        }
      },
      "MoveData": {
        "type": "object",
        "properties": {
          "from_square": {"type": "string"},
          "to_square": {"type": "string"},
          "san": {"type": "string"},
          "piece_type": {"type": "string"}
        }
      },
      "MoraleChange": {
        "type": "object",
        "properties": {
          "piece_id": {"type": "string"},
          "event_type": {"type": "string"},
          "change": {"type": "integer"},
          "morale_after": {"type": "integer"},
          "description": {"type": "string"}
        }
      },
      "MoveAnalysis": {
        "type": "object",
        "properties": {
          "move_quality": {"type": "integer", "minimum": 1, "maximum": 10},
          "evaluation": {"type": "number"},
          "threats": {"type": "array", "items": {"type": "string"}},
          "opportunities": {"type": "array", "items": {"type": "string"}},
          "analysis_text": {"type": "string"}
        }
      },
      "PersuasionRequest": {
        "type": "object",
        "required": ["piece_id", "target_square", "argument"],
        "properties": {
          "piece_id": {"type": "string"},
          "target_square": {"type": "string", "pattern": "^[a-h][1-8]$"},
          "argument": {"type": "string", "maxLength": 500},
          "is_voice": {"type": "boolean", "default": false}
        }
      },
      "PersuasionResponse": {
        "type": "object",
        "properties": {
          "success": {"type": "boolean"},
          "probability": {"type": "number"},
          "piece_response": {"type": "string"},
          "move_executed": {"type": "boolean"},
          "board_state": {"$ref": "#/components/schemas/BoardState"},
          "evaluation": {"$ref": "#/components/schemas/PersuasionEvaluation"}
        }
      },
      "PersuasionEvaluation": {
        "type": "object",
        "properties": {
          "logic_score": {"type": "integer"},
          "personality_match": {"type": "integer"},
          "morale_modifier": {"type": "integer"},
          "trust_modifier": {"type": "integer"},
          "urgency_factor": {"type": "integer"},
          "total_probability": {"type": "number"}
        }
      },
...
      "ChatMessage": {
        "type": "object",
        "properties": {
          "id": {"type": "string"},
          "message_type": {"type": "string"},
          "sender_name": {"type": "string"},
          "content": {"type": "string"},
          "metadata": {"type": "object"},
          "created_at": {"type": "string", "format": "date-time"}
        }
      },
      "SendMessageRequest": {
        "type": "object",
        "required": ["content"],
        "properties": {
          "content": {"type": "string", "maxLength": 1000},
          "message_type": {"type": "string", "default": "player_message"}
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "sessionAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Session-ID"
      }
    }
  }
}